chore: house keeping refactoring and cleanups
