+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,18 +1,13 @@
 # AniTrend Pull Request
 
-Thank you for contributing! Please take a moment to review our [**contributing
-guidelines**](https://github.com/AniTrend/android-emojify/blob/develop/CONTRIBUTING.md)
-to make the process easy and effective for everyone involved.
+Thank you for contributing! Please take a moment to review our [**contributing guidelines**](https://github.com/AniTrend/android-emojify/blob/develop/CONTRIBUTING.md) to make the process easy and effective for everyone involved.
 
-**Please open an issue** before embarking on any significant pull request, especially those that
-add a new library or change existing tests, otherwise you risk spending a lot of time working
-on something that might not end up being merged into the project.
+**Please open an issue** before embarking on any significant pull request, especially those that add a new library or change existing tests, otherwise you risk spending a lot of time working on something that might not end up being merged into the project.
 
 Before opening a pull request, please ensure you've done the following:
 <!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
 
-- You have followed our [**contributing
-  guidelines**](https://github.com/AniTrend/android-emojify/blob/develop/CONTRIBUTING.md)
+- You have followed our [**contributing guidelines**](https://github.com/AniTrend/android-emojify/blob/develop/CONTRIBUTING.md)
 - Double checked that your branch is based on `develop` and targets `develop` (where applicable)
 - Pull request has tests (if applicable)
 - Documentation is updated (if necessary)
@@ -32,6 +27,4 @@ Before opening a pull request, please ensure you've done the following:
 
 <!--- Be kind to code reviewers, and please try to keep pull requests as small and focused as possible :) -->
 
-**IMPORTANT**: By submitting a patch, you agree to allow the project
-owners to license your work under the terms of
-the [Apache License 2.0](https://github.com/AniTrend/android-emojify/blob/develop/LICENSE.md).
+**IMPORTANT**: By submitting a patch, you agree to allow the project owners to license your work under the terms of the [Apache License 2.0](https://github.com/AniTrend/android-emojify/blob/develop/LICENSE.md).
