+++ app/src/main/java/io/wax911/emojifysample/EmojiInitializer.kt
@@ -2,7 +2,7 @@ package io.wax911.emojifysample
 
 import io.wax911.emojify.contract.serializer.IEmojiDeserializer
 import io.wax911.emojify.initializer.AbstractEmojiInitializer
-import io.wax911.emojify.serializer.KotlinxDeserializer
+import io.wax911.emojify.serializer.kotlinx.KotlinxDeserializer
 
 class EmojiInitializer : AbstractEmojiInitializer() {
     override val serializer: IEmojiDeserializer = KotlinxDeserializer()
