+++ buildSrc/src/main/java/io/wax911/emoji/buildSrc/Libraries.kt
@@ -8,7 +8,7 @@ object Libraries {
         object Emojify {
             val emojify = Modules.Library.Emojify.path()
             val contract = Modules.Library.Contract.path()
-            val serializerKotlinx = Modules.Library.SerializerKotlinx.path()
+            val serializerKotlinx = Modules.Library.SerializerKotlinX.path()
             val serializerGson = Modules.Library.SerializerGson.path()
             val serializerMoshi = Modules.Library.SerializerMoshi.path()
         }
