+++ buildSrc/src/main/java/io/wax911/emoji/buildSrc/module/Modules.kt
@@ -18,7 +18,7 @@ internal object Modules {
     enum class Library(override val id: String) : Module {
         Emojify("emojify"),
         Contract("contract"),
-        SerializerKotlinx("serializer:kotlinx"),
+        SerializerKotlinX("serializer:kotlinx"),
         SerializerGson("serializer:gson"),
         SerializerMoshi("serializer:moshi"),
     }
