+++ buildSrc/src/main/java/io/wax911/emoji/buildSrc/plugin/CorePlugin.kt
@@ -13,8 +13,7 @@ open class CorePlugin : Plugin<Project> {
     /**
      * Inspecting available extensions
      */
-    @Suppress("UnstableApiUsage")
-    internal fun Project.availableExtensions() {
+    private fun Project.availableExtensions() {
         val extensionSchema = project.extensions.extensionsSchema
         extensionSchema.forEach {
             println("Available extension for module ${project.path}: ${it.name} -> ${it.publicType}")
@@ -24,8 +23,7 @@ open class CorePlugin : Plugin<Project> {
     /**
      * Inspecting available components
      */
-    @Suppress("UnstableApiUsage")
-    internal fun Project.availableComponents() {
+    private fun Project.availableComponents() {
         val collectionSchema = project.components.asMap
         collectionSchema.forEach {
             println("Available component for module ${project.path}: ${it.key} -> ${it.value}")
