+++ buildSrc/src/main/java/io/wax911/emoji/buildSrc/plugin/components/AndroidOptions.kt
@@ -26,7 +26,7 @@ private fun Project.createMavenPublicationUsing(sourcesJar: Jar) {
             version = props[PropertyTypes.VERSION]
 
             artifact(sourcesJar)
-            artifact("${project.buildDir}/outputs/aar/${project.name}-release.aar")
+            artifact("${project.layout.buildDirectory}/outputs/aar/${project.name}-release.aar")
             from(component)
 
             pom {
@@ -148,7 +148,6 @@ private fun Project.createDokkaTaskProvider() = tasks.named<DokkaTask>("dokkaHtm
 }
 
 
-@Suppress("UnstableApiUsage")
 internal fun Project.configureOptions() {
     if (isLibraryModule()) {
         println("Applying additional tasks options for dokka and javadoc on ${project.path}")
@@ -161,7 +160,7 @@ internal fun Project.configureOptions() {
         }
 
         val classesJar by tasks.register("classesJar", Jar::class.java) {
-            from("${project.buildDir}/intermediates/classes/release")
+            from("${project.layout.buildDirectory}/intermediates/classes/release")
         }
 
         artifacts {
